DABOM-322 feat: 컴포넌트 분리
